+++ graphql-strava-activities.php
@@ -110,6 +110,7 @@ function wpgraphql_strava_init(): void {
 	require_once WPGRAPHQL_STRAVA_DIR . 'includes/admin.php';
 	require_once WPGRAPHQL_STRAVA_DIR . 'includes/class-wpgraphql-strava-activities-list-table.php';
 	require_once WPGRAPHQL_STRAVA_DIR . 'includes/graphql.php';
+	require_once WPGRAPHQL_STRAVA_DIR . 'includes/rest-api.php';
 	require_once WPGRAPHQL_STRAVA_DIR . 'includes/shortcodes.php';
 }